+++ .gitignore
@@ -1,0 +1,4 @@
+.vscode/tasks.json
+.vscode/launch.json
+main-file.exe
+main-console.exe
