+++ kursiokai.txt
@@ -1,4 +1,4 @@
 Pavarde     Vardas    Egzaminas    ND1  ND2   ND3   ND4  ND4  ND5
-Pavarde1    Vardas1   8            n    5     6     n    10   4
-Pavarde2    Vardas2   6            5    6     O     8    2    5
-Pavarde3    Vardas3   10           9    10    10    5    n    6
+Grilas      Vardas1   2            n    5     6     n    10   4
+Wixed       Vardas2   3            5    6     O     8    2    5
+Aruozas     Vardas3   10           5   10    10    10   n   10
